// src/models/dog.model.ts
export interface Dog {
    id: string;
    name: string;
    sex: string;
    size: string;
    breed: string;
    age: string;
    color: string;
    fplace: string;
    found: string;
    behaviourSenior: string,
    behaviourKids: string;
    behaviourDogs: string;
    behaviourCats: string;
    description: string;
    collar: string;
    coat: string;
    status: string;
    profilepic: string;
    idDogPound: string;
    adoptStatus: string;
    adoptees: string[];
}